Add release workflow
This adds a workflow that can create a PR transitioning a branch from
in development to release, or vice versa. A releaser:

1. executes the manual action to create a PR updating the maven versions
2. merges this PR after CI runes
3. manually tags and pushes up the release tag, triggering creation of a
   tagged image
4. executes the transition action to create a PR updating the branch for
   development
5. merges this PR

Images will be created on every push to main and branches named
`release/**`. The image will be tagged with the branch name.

Images will be created when any tag starting with `v` is pushed. The
image will be tagged with the git tag.

Nothing is tagged as latest currently.

Signed-off-by: Robert Young <robeyoun@redhat.com>